First Version of code
Haven't done Exception Handling etc, Main Goal for this Task is to make things work. There are PNG files for each Record attached as well. Test output file is attached as well.

One thing which is pending is (FreeLancer and AuthorName in the JSON are not populated). Have to look into that